matrix

+++ LAB/ncr.c
@@ -5,13 +5,13 @@ int fact(int num);
 int main(){
     
     int n, r;
-    int ncr;
+    float ncr;
     printf("Enter the value of n: ");
     scanf("%d", &n);
     printf("Enter the value of r: ");
     scanf("%d", &r);
     ncr = fact(n) / (fact(r) * fact(n - r));
-    printf("Value of %dC%d is: %d", n, r, ncr);
+    printf("Value of %dC%d is: %0.2f", n, r, ncr);
 }
 
 int fact(int num){
